+++ example_apps/plot_demo.py
@@ -1,9 +1,30 @@
 import tkinter
 from tk_builder.panels.pyplot_panel import PyplotPanel
-from example_apps.plot_demo.button_panel import ButtonPanel
 from tk_builder.panel_builder import WidgetPanel
 from tk_builder.widgets import widget_descriptors
 import numpy as np
+
+
+class ButtonPanel(WidgetPanel):
+    """
+    Basic button panel.
+    """
+    _widget_list = ("single_plot", "multi_plot", "animated_plot")
+    single_plot = widget_descriptors.ButtonDescriptor("single_plot")
+    multi_plot = widget_descriptors.ButtonDescriptor("multi_plot")
+    animated_plot = widget_descriptors.ButtonDescriptor("animated_plot")
+
+    def __init__(self, parent):
+        """
+
+        Parameters
+        ----------
+        parent
+            The parent widget.
+        """
+
+        WidgetPanel.__init__(self, parent)
+        self.init_w_horizontal_layout()
 
 
 class PlotDemo(WidgetPanel):
